Update test_harness.py logout test
The DOCUMENT_NAME file was deleted on the server in the prior delete tests (line 333), so it should no longer exist on the file. However, user 2 still has the file checked out (line 283). Upon logout, this file should be checked back in by user2 and the owner set to user2. 
quote> 
quote> feel free to reject this PR if you disagree.